Rename app.js to rest-api.js (#57)
