+++ src/components/parameters.js
@@ -28,7 +28,7 @@ import Typography from '@material-ui/core/Typography';
 import Switch from '@material-ui/core/Switch';
 
 import { DARK_THEME, LIGHT_THEME } from '../redux/actions';
-import { updateConfigParameter } from '../utils/api';
+import { updateConfigParameter } from '../utils/rest-api';
 import {
     PARAMS_THEME_KEY,
     PARAMS_TIMELINE_DIAGONAL_LABELS,
